refactor: migrated the saveImageInfo logic into a separate helper file
